Simplify comment labeling in image.php
The image template is the only one to label "comments and trackbacks are open" "comments are closed but trackbacks are open" etc. ad nauseam. Few people are using those options any more, and it's a lot of clutter on a template that (if anything) should be simpler and cleaner than all the rest - we should delete all that and match the comment presentation to the way it's done in page.php and single.php