Simplify Clearing
(Second opinion needed)

This should simplify clearing and make the development faster. 

Current clearing causes the issues with floats and display properties in header and footer area with default layout and even more if extended.

The *site which is being cleared 4 times in the header.

<pre>
<code>
.site-header
.site-branding
.site-title
.site-content
</code>
</pre>

And one more time for the *content ID.

<pre>
<code>
#content
</code>
</pre>

<pre><code>:before</code></pre> and <pre><code>:after</code></pre> already exist and will produce the same "affect" from the inner area and that is <pre><core>#content</code></pre>.

.clear remains just in case if needed.

Tested on several sites for my clients and haven't found any issues. Browser support IE8+ just like the original one.