An install bash script that auto-customizes _s
Inspired by the oh-my-zsh install script[1], this script depends on
bash, git, and Perl. It's really quite simple: it prompts the user for a
theme name, downloads the theme from GitHub using git, then does a Perl-powered
search-and-replace as the README instructs the user to do manually. Oh,
it also auto-generates a new README file.

Disclaimer: I'm not typically a bash scripter, so some of my code will
need some improving. But, I've confirmed that it at least works by going
through the code manually and looking for anything weird due to the
search-and-replace, and I briefly tested it on a remote hosted WP
instance.

[1]:
https://github.com/robbyrussell/oh-my-zsh/blob/master/tools/install.sh