Forms
Proposing to reword Inputs and Buttons with Forms. This way we can cover more than just couple of elements. Select feels like an outsider and textarea too, even though it's technically an input.

This will also be a perfect match with the rest :)

N.B. Something went wrong with my last proposal, sorry about that.