_s: Encourage the use of sidebar descriptions.
Sidebar descriptions provide an opportunity to provide context about
where the sidebar is located or when it is loaded. We should encourage
theme developers to use them, similar to how we hint default images for
custom headers and backgrounds.

The callback’s outdated documentation describes the function in an
early state of Toolbox (introduced in r3678-wpcom-themes), the earliest
public record would be
https://themes.trac.wordpress.org/browser/toolbox/0.3/functions.php#L99
This is as good an opportunity to update it as any.

Props @emiluzelac for initial patch.
Fixes #433.