widgettitle value
```widgettitle``` doesn't really have any value here or in core. Proposing to change to ```widget-title``` instead and maybe add that into stylesheet too. 

The class is available only in /widgets.php and not present in any of the stylesheet's, at least I could not locate one. As in my previous proposal, this is for the consistency reasons only.