_s: Simplify entry meta output
- Fix entry-footer inconsistency between content-*.php files
- Move in a function <code>_s_entry_meta()</code> to avoid duplication

---

At the moment there is an inconsistency between content.php and
content-single.php regarding the .entry-footer output.

In my opinion content.php output is would be better to use because it's
easier to translate and allow more customisation.

Finally by moving it into <code>inc/template-tags.php</code> we are
getting rid of some duplicated code.