Normalize rather than Reset.
The reset we used was a mix of Normalize, the Paul Irish reset,
sprinkled with some blueprint.css, and base styles. It was hard to
maintain and just seemed outdated (setting the base font-size to 10px
instead of 16px, among other things).

Normalize sets saner defaults and is generally just not as disruptive
as a full reset.

The code was added as is, has only its comments stripped (for now), and
is not mixed with any styles to make future updates to it as easy as
possible.

See #3, #44, #174, #267, #617.