Improve translation string of _s_posted_on()
The output of _s_posted_on() function is split in 2 translation strings with no possibility to interchange them. Currently the output is locked to "Posted on {date} by {author}" or even "{author} wrote this on {date}". Combining them into a single translation string allows for output like "Posted by {author} on {date}" and it even improves the translation possibilities for languages with non-latin semantics.

I know that I have made some dramatic changes to established translation strings, but also adding HTML tags to translation strings isn't the most elegant solution and this pull request doesn't need to be merged as is. I propose a discussion wether prepositions like "on" and "by" actually make a semantic sense when wrapped in <span> tags together with the meta and wether the word "Posted" is directly related to the post date or makes sense even outside this context.