Move aside from the widgets to the entire sidebar.
It appears that the entire sidebar should be one large aside instead of
making each widget it's own aside. At least that's how I understand it
when reading w3. It would make sense to make each widget it's own aside
if not all of the widgets were asides, but otherwise the entire sidebar
should be considered one aside.

Plus I think this would make more sense for accessibility and with the
widgets titles being h2 because if the entire sidebar had a title that
would be the h1.

https://www.w3.org/wiki/HTML/Elements/aside