Adding _s_queue_transient_flusher
Queue the add_action for save_post_{post_type} only if the post type
has the 'category' taxonomy. Hooking late on init to let plugins
register custom posts types first and so the action added in the admin
and on the front-end in case a front end editor is being used to save
the post.