Enqueue Font Awesome for use
This does not add to the core functionality of _S theme but I think it is becoming a "necessary evil". Since DashIcons is not increasing the number of growing social media icons and several others, developers feel like they need to have the freedom to express their work with more icons. Enqueueing Font Awesome for use in WP Admin and Template would help a lot to solve this issue.