Remove automated number version from theme
I have noticed that  several users of WordPress do not update their core
on live production even when prompted too in the dashboard. Some
security folks argue that this could be a problem when hackers use the
version vulnerabilities for the "un-updated" core to hack the website.
This information can easily be got from the source in the web browser

Ben Sibley argues that for the meta tag generator versioning - "serves
little purpose, but potentially is/could be used for gathering
statistics by identifying sites as using WordPress. Some have said it is
a potential security concern because it displays the version of WP
you’re using. I’m no security expert, but I’m sure that’s an invalid
claim or it wouldn’t be in core since V2.5."

Ref:
https://www.competethemes.com/blog/wordpress-meta-tags-ultimate-guide/

I would love a second opinion on this. Thus the comment out in the
functions.php file. If this workable, a merge would be good.