Update header.php
Using And && instead of OR ||, the h1 tag will be used only if we select to display our latest posts in the front page. If we choose a static page as our front page, the h1 tag will never be used.
Using OR ||, the h1 tag will be used in the front page and in the posts page.