Travis: speed up build times by disabling Xdebug
Xdebug is only **needed** when creating code coverage reports. Otherwise, it's just a nice extra, but a slow extra.
Disabling it will speed up the build times.

As suggested by John Blackbourn in https://johnblackbourn.com/reducing-travis-ci-build-times-for-wordpress-projects/

With an additional improvement to stabilize the command as suggested by Niklas Keller in  https://twitter.com/kelunik/status/954242454676475904