QA: Break stylesheets out of browser cache on each new release of a theme
The `wp_enqueue_style()` function allows for passing a `$version` parameter which is used to force browsers to load a fresh copy of a stylesheet when it has been changed.

If this parameter is not passed, the version number will be set to the version of the WP install which doesn't make sense at all for a theme which is not shipped with WP itself.

This commit adds a `_S_VERSION` constant which should be changed to the current version number of the actual finished theme on each new release.

This constant is subsequently used in all `wp_enqueue_style()` function calls to ensure that on each new release of the theme, the browser cache is cleared for these stylesheets.

Ref: https://developer.wordpress.org/reference/functions/wp_enqueue_style/