:rocket: Build: use PHPCompatibilityWP
Since mid July, the PHPCompatibility project offers dedicated CMS rulesets which already account for all back-fills/poly-fills.

This commit removes the `_s` native version of that ruleset in favour of using this dedicated WP-based ruleset.

Additionally, the PHPCompatibility repo has moved to a dedicated GitHub organisation, so that change is included here as well.

Refs:
* https://github.com/PHPCompatibility/PHPCompatibility/releases/tag/8.2.0
* https://github.com/phpcompatibility/phpcompatibilitywp

Note: I would strongly recommend installing these dependencies via Composer instead. grappler has created a good setup for this in PR 1291 which I would recommend for merging, though it will need to be updated to be in line with the changes made in this PR.