Add option for excerpt
Underscores only supports the use of "read more" inside of the content. But I think enough devs would like to have the option of the "classic" excerpt from the start. Also for some clients might be easier to understand/learn the use of an excerpt that is outside of the content.